CSCMETAX-61: [REF] Create tests/api/base/apitests/common as a place for common api functionality testing
Moved some tests from CatalogRecord tests to common: PUT if-unmodified-since tests,
GET if-modified-since, paging.
