+++ src/metax_api/tests/api/base/apitests/common/__init__.py
@@ -1,0 +1,2 @@
+from .read import *
+from .write import *
